Add TBC DASH compatible unit test example
DASH provides an alternative framework to unit-test ShaCL "natively"
with RDF, via for e.g. DASH-compliant ShaCL processors in editors like
TopBraid Composer (TBC). Here we provide such a test for TBC Community
Edition (CE) 6+, as TopBraid no longer provides a CE for their flagship
editor product. While PySHACL supports DASH, the test files will be
different as the expected results are different in TBC's implementation.

TopBraid/DASH unit tests are not very vendor-agnostic at this time, as
far as tools external to the product/company are concerned. This is
because tests are done against a local graph, and any rules pertaining
to standard vocabularies -- as is the case for us with OWL and ShaCL
profiles -- will throw off the tests as the result contains A LOT of
triples from that local graph, containing resources from such standard
vocabs.

The TBC DASH tests can be generated by simply concatenating, on top of
an editor-specific preamble section that includes prefixes and imports,
the example data and the corresponding shapes into a `*.test.ttl`,
opening it up with TBC, running all tests, and then choosing to "create
a test case". This will modify the file in-place and add the expected
triples, which can be too numerous due to the above. Therefore, we limit
our examples at this time :)

In general, we might make the below comparison between DASH and
a pure Python-based SHACL test suite.

Pros of DASH:

- Granular specification of result by default (conforms and conforms
how)

	- In Python, have to extract result components and assert

- Standard approach

	- In Python, freedom of expression means people can do things
different ways

Cons:

- Rigid organization of test suite components (same graph has test
cases, shapes, data)

	- In python, anything goes (look here for shapes, there for
data, etc.); `owl:import`s cannot deal with relative paths

- Only one test data instance per case (focus node)

	- In Python, anything goes (test data can have multiple
instances that do or do not conform; can also be a pro depending on
perspective)

- Uses implicit class targets, so lower separation of concerns

	- In Python, anything goes (explicit targets can be used)

- Rigid environment expectations require different rules for special
cases (instances in own graph which can be problematic for shapes on
standard vocabs OWL and SHACL)