Add partial lexicalisation tests
While SEMIC rule SC-R2 discusses recommended lexicalisation properties
that constitute a _lightweight_ profile, it is not possible to "close"
the shapes to restrict usage of properties to only those listed, as it
would be too restrictive for practical use.

There is an overlap of SC-R2 with SC-R3, which elaborates on recommended
lexicalisation properties from a set of two vocabularies (`rdfs` and
`skos`). Here, again, the restrictions would be impractical, and the
shapes too complex.

However, while not directly prescribed or specified by the rules, we can
derive two other rules from them:

1. There MUST only ever be at most one label value per language and
without, for the two primary lexicalisation properties `rdfs:label` and
`skos:prefLabel` (we may also later expand to `sh:name` for ShaCL)

2. There SHOULD always be at least one of the two primary lexicalisation
properties expressed for a given OWL resource (again, this will be
different for ShACL)

The former involves complex shapes requiring a "qualified value shape".
Together, they ensure that best practices are maintained with regards to
usage of lexicalisation properties (exactly one occurence of each in a
language or plain).

SEMIC rules: SC-R1 (partial), SC-R2 (partial)

Co-authored-by: Ahmad, Jana, Ing <ahmadjan@fel.cvut.cz>