Add subsumption test
Any relationship that holds between classes can only be the subsumption
`rdfs:subClassOf`, and that between properties (of any of the two OWL
types) can only be the subsumption `rdfs:subProperty`. These can only be
expressed by custom SHACL-SPARQL queries.

Note that the `rdf:type` property must also be considered together as
part of the subsumption (and also because the conventions allow and
expect such declarations, even if implicitly).

SEMIC rules: SC-R2