Add OWL ontology test
While there is no specific signature of an OWL ontology other than its
resources (specified by other rules already), and OWL1 vs. OWL2 is not
relevant as SEMIC doesn't even recommend OWL2 features, we can still
check for an OWL ontology header.

We derive two shapes for the purpose:

- A node shape targeting a dummy node with a custom SPARQL query that
checks if there does NOT exist a statement with `rdf:type` (not mandated
but good to have as this is always an implied assumption)

- A node shape targeting the class `owl:Ontology` with a restriction of
minCount=1 on the inverse of the path `rdf:type` (there must exist at
least one instance of the target class)

SEMIC rules: SC-R1