Add some performance indications for SHACL-SPARQL
Profiling (manually) against ePO-core revealed some important
performance implications of using custom SHACL queries in the shapes. In
total, approximately 70% of the performance impact is from these
queries. We add the deactivated flag for convenience of marking `true`
if running these shapes are undesired due to this. We should probably
rethink or implement a mechanism to read an exclusion list.