Make SC-R2 domain-range rule a Warning
Apparently this makes more sense as a warning due to the use of
`no..shall` in the language. This, however, may not be consistent
throughout the conventions. Furthermore, another (sub)rule raises a
violation for use of non-subsumption relationships, which non-compliant
data for this one will also trigger, so that's another
conflict/inconsistency.