Add SHACL severity declaration test
We can only test whether a property shape has a severity declared,
otherwise we would most likely have to employ SPARQL to cater for the
conditional case or dependence where a severity on a node shape is
required only when there is no associated property shape. There was not
a better way found among the SHACL Advanced or DASH specs. Because such
a situation is rare in practice, we choose to KISS.

SEMIC rules: DSC-R5