Improve severity declaration test with XONE
Handle the situation where either the node or property shape should have
a severity, but the former only when there is no related latter shape.

This clever trick with XONE (exactly or "exclusively one of") also works
even if the PropertyShape is not declared when typically related through
a blank node (although whether this situation is allowed is another
topic which another rule/test currently serves) owing to the use of
SequencePath.

This leaves out "freestanding" property shapes, but whether we should
tackle them is not yet clear.