Add select few expressions test
We interpret the rule as recommending property expressions to be limited
to a select few, rather than recommending a minimum required set
(membership). However, the individual properties can be set to `minCount
1` to enable recommendations of those properties if such a minimum is
required.

This is the first use of a "closed" shape, which does not permit
properties other than those in the ignore list (which we currently
populate with the primary OWL object relations).

SEMIC rules: SC-R2

Co-authored-by: Ahmad, Jana, Ing <ahmadjan@fel.cvut.cz>