Add semantic versioning test, sans versionIRI
We interpret the rule as alluding to (the required presence of)
`owl:versionInfo` when it prescribes a format and not (also)
`owl:versionIRI`. The IRI should ideally match the version in one form
or another (in any part of it) but that would be a more complex shape or
regular expression to validate. One place on the Web where both
properties are treated is at
https://obofoundry.org/principles/fp-004-versioning.html.

SEMIC rules: PC-R3