Add minimum ontology header test
This is a derived subrule as the conventions do not explicitly ask for a
minimum recommended set of properties, but rather ask to be limited to a
few of them. We combine those with common usage for e.g. by ePO in order
to make some recommendations.

SEMIC rules: SC-R2 (partial)

P.S.: Language-tagged strings are different from plain strings and the
union value type cannot be checked. There is `dash:StringOrLangString`
but PySHACL appears to not have support for the DASH vocabulary. If we
were to make a DASH test case, however, it might've been validated with
`check_dash_result=True`.

Co-authored-by: Ahmad, Jana, Ing <ahmadjan@fel.cvut.cz>