Add advanced definitions test
OWL restrictions are the key features of OWL2, which are typically
expressed through a blank node (bNode) typed as `a owl:Restriction`. For
these, we simply restrict usage of subsumption relationships to IRIs
and, for cases where named individuals are used and not bNodes, the
SC-R2 subrule/test `only_owl_resources` will raise a violation anyway
for types beyond the basics.

For the same reason (forbidden types), declaring property
characteristics (e.g. `a owl:TransitiveProperty`), or properties that
can only be used in a restriction (e.g. property constraints such as
`owl:someValuesFrom` and `owl:cardinality`) will raise the SC-R2
violation. For everything else that qualifies as "an advanced property",
be it whether used on a class or object property, we simply use a SHACL
alternative list and restrict cardinality to zero.

SEMIC rules: SC-R2 (partial)