Fix SPARQL-query shape for ITB/Java-SHACL
The ITB Java SHACL processor flags this particular SPARQL query due to a
missing binding for `?this` (`$this` specifically in this context but
both syntaxes appear to work), which the query does not use and did not
return. It appears simply returning it suffices, even though it has no
value.