ITB: Validate OWL and SHACL separately
Use the latest version of OWL and SHACL shapes sufficient for testing
real-world files (use the non-ePO example data in `examples/` if you
don't get results for ePO).

Note that combined OWL/SHACL files and their validation will work but
will either (a) not make sense or (b) not be able to produce output or
export citing too many results. Therefore, we say this support is
incomplete or partial, until such a time as combined validation is
handled properly (there are some nuances, shapes that should be
shared/common, etc.)